+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.0.1
+
+* change handleSpecialText to hasSpecialInlineSpanBase(extended_text_library)
+* add hasPlaceholderSpan(extended_text_library)
 ## 5.0.0
 
 * Merge from Flutter v1.22